requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# verify/bot.py: 5
async_timeout == 3.0.1

# verify/bot.py: 2,6
discord.py == 1.7.3
